Fix license in pyproject.toml
The format for the license field changed since PEP 639; it is now
required to provide the license identifier in `license`, and the list of
actual files (glob patterns) in `license-files`.

We currently have only one file for the license, but that may change if
at some point we include data or third-party source code (other than
libs). The glob pattern should cover most usages, without requiring to
change the pyproject.toml again.