Fix dependencies in docs CI/CD
We need to install the source code dependencies to build the docs. The CI/CD used a requirements.txt file for the *docs* dependencies, but we do not have such a file for the source code dependencies.
This commit installs dependencies through the pyproject.toml, in the same way as the "check" CI/CD (except that we install the "docs" extras rather than "dev").