Revert the flower's planted_by when creating grid and allow None
Since a previous commit, the flowers' now track which agent planted which
is a good thing, because it allows to accurately track the number of
plants harvested and planted by each agent.

However, this requires flowers to know which agent planted them; when an
agent performs the action of planting, this is easy to do, as we can
pass the agent to the Flower constructor.
When a flower is initially planted (i.e., when the grid world is first
created), this is a bit more complex. The previous commit changed the
way the grid world must be defined, to specify the "planter agent" for
each flower.

Flowers that are initially planted should not be considered to have been
planted by a specific agent, because agents have not done anything at
this point. They are given an initial state (which, metaphorically,
results from the nature), and must do as best as possible with this
state.

Thus, this commit reverts the obligation to specify agents when
initializing the grid world; instead, we can specify None.

To summarize:
- flowers planted by agents (during an action) must specify the planter
  agent;
- flowers planted during the world initializing must specify None as the
  planter agent.