Add helper script to submit slurm jobs
This helper script relies on the config files defined in previous
commits. While the config files are required, the helper script is
only meant to facilitate and document how to submit jobs to SLURM
using the config files. That is, you can manually type the command
from the helper script in your shell, change a few variables, and
it will work.